pyproject.toml
[tool.poetry]
name = "graphtest"
version = "0.1.0"
description = ""
authors = ["ahMADASSadi <ds1380ahmad@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django-graphql-jwt = "^0.4.0"
django = "^5.1.6"
graphene-django = "^3.2.2"
django-filter = "^25.1"
pillow = "^11.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
